Add files via upload
Developement of machine learning algorithm for detection of breast cancer stages using advanced image processing techniques on
available medical datasets and classifying the images based on cancer stages. It's deployment on the GUI application its helps for
real time detection and prediction of breast cancer staging directly on the device, with the need of cloud connectivity, speed, and
reliability in medical diagnostics.